wip refactor hap functions
