+++ malariagen_data/ag3.py
@@ -40,7 +40,6 @@ SITE_ANNOTATIONS_ZARR_PATH = (
 )
 DEFAULT_MAX_COVERAGE_VARIANCE = 0.2
 PCA_RESULTS_CACHE_NAME = "ag3_pca_v1"
-SNP_ALLELE_COUNTS_CACHE_NAME = "ag3_snp_allele_counts_v2"
 FST_GWSS_CACHE_NAME = "ag3_fst_gwss_v1"
 H12_CALIBRATION_CACHE_NAME = "ag3_h12_calibration_v1"
 H12_GWSS_CACHE_NAME = "ag3_h12_gwss_v1"
@@ -109,7 +108,6 @@ class Ag3(AnophelesDataResource):
     virtual_contigs = "2RL", "3RL"
     _site_annotations_zarr_path = SITE_ANNOTATIONS_ZARR_PATH
     _pca_results_cache_name = PCA_RESULTS_CACHE_NAME
-    _snp_allele_counts_results_cache_name = SNP_ALLELE_COUNTS_CACHE_NAME
     _fst_gwss_results_cache_name = FST_GWSS_CACHE_NAME
     _h12_calibration_cache_name = H12_CALIBRATION_CACHE_NAME
     _h12_gwss_cache_name = H12_GWSS_CACHE_NAME
@@ -117,8 +115,6 @@ class Ag3(AnophelesDataResource):
     _g123_calibration_cache_name = G123_CALIBRATION_CACHE_NAME
     _h1x_gwss_cache_name = H1X_GWSS_CACHE_NAME
     _ihs_gwss_cache_name = IHS_GWSS_CACHE_NAME
-    phasing_analysis_ids = ("gamb_colu_arab", "gamb_colu", "arab")
-    _default_phasing_analysis = "gamb_colu_arab"
 
     def __init__(
         self,
@@ -150,6 +146,7 @@ class Ag3(AnophelesDataResource):
             },
             site_filters_analysis=site_filters_analysis,
             default_site_mask="gamb_colu_arab",
+            default_phasing_analysis="gamb_colu_arab",
             bokeh_output_notebook=bokeh_output_notebook,
             results_cache=results_cache,
             log=log,
